# proc: cache result of GetG